feat(my-account): apply Newspack UI styles to My Account w/ env constant (#3951)
* refactor: reorganize My Account files and code for modularity

* feat: apply Newspack UI styles to My Account with env constant

* feat: add variables module and use custom theme accent/contrast colors

* style: change inline table buttons to x-small

* style: more controlled layout margins/padding for mobile

* fix: no flex without nav sidebar

* fix: apply box styles to messages

* fix: use notice styles, not box

* fix: remove unnecessary untils module

* style: hover background color for nav buttons

* refactor: move woocommerce files to plugins folder

* refactor: no need to extend base class

* chore: remove unused `use`

* fix: line-height

* style: update nav buttons to match --ghost styles

* style: fix button styles in subscription tables

* style: standardize spacing for elements after a .shop_table

* style: make .delete buttons destructive

---------
