+++ includes/plugins/woocommerce/class-woocommerce-cli.php
