+++ includes/plugins/woocommerce/class-woocommerce-connection.php
