+++ includes/plugins/woocommerce/class-woocommerce-cover-fees.php
