+++ includes/plugins/woocommerce/class-woocommerce-logs.php
