+++ includes/plugins/woocommerce/class-woocommerce-order-utm.php
