+++ includes/plugins/woocommerce/class-woocommerce-products.php
