+++ includes/plugins/woocommerce/my-account/templates/myaccount-after-delete-account.php
@@ -21,4 +21,3 @@ defined( 'ABSPATH' ) || exit;
 		<?php echo esc_html__( 'Return to home page', 'newspack-plugin' ); ?>
 	</a>
 </div>
-
