+++ includes/plugins/woocommerce/my-account/templates/myaccount-delete-account.php
