+++ includes/plugins/woocommerce/my-account/templates/myaccount-verify.php
