+++ src/my-account/v0/index.js
@@ -1,0 +1,8 @@
+/**
+ * Customizations for Newspack My Account pages: v0.x.x.
+ */
+
+/**
+ * Internal dependencies.
+ */
+import './style.scss';
