+++ src/my-account/v1/_content.scss
@@ -1,0 +1,16 @@
+/**
+ * Main content area
+ */
+
+.woocommerce-account.newspack-my-account {
+	.woocommerce-MyAccount-content {
+		padding-left: var(--newspack-ui-spacer-5);
+		padding-right: var(--newspack-ui-spacer-5);
+		@media ( min-width: 782px ) {
+			margin-left: auto;
+			margin-right: auto;
+			max-width: 782px;
+			padding: 0;
+		}
+	}
+}
