+++ src/my-account/v1/_forms.scss
@@ -1,0 +1,12 @@
+/**
+ * Forms
+ */
+
+.woocommerce-account.newspack-my-account {
+	p.woocommerce-form-row,
+	.woocommerce-form-row p {
+		color: var(--newspack-ui-color-neutral-60);
+		font-size: var(--newspack-ui-font-size-xs);
+		line-height: var(--newspack-ui-line-height-xs);
+	}
+}
