+++ src/my-account/v1/index.js
@@ -1,0 +1,8 @@
+/**
+ * Customizations for Newspack My Account pages: v1.x.x.
+ */
+
+/**
+ * Internal dependencies.
+ */
+import './style.scss';
