+++ src/newspack-ui/scss/elements/woocommerce/_index.scss
@@ -1,1 +1,2 @@
 @use "overrides";
+@use "my-account";
